requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
python-docx==0.8.11
requests==2.28.1
selenium==4.6.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
